server/requirements.txt
click==8.0.3
fastapi==0.70.0
passlib==1.7.4
pydantic[email,dotenv]==1.8.2
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.5
sentencepiece==0.1.96
slugify==0.0.1
spacy==3.2.3
srsly==2.4.2
supervisor==4.2.4
uvicorn==0.16.0
websocket-client==1.3.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
git+https://oauth:${GITLAB_OAUTH_TOKEN}@storfa.techiaith.cymru/cyfieithu/techiaith-utils.git@v2022.03.1
